+++ api/deploys.py
@@ -42,7 +42,7 @@ def trigger_a_deploy():
         print(e)
         return False
 
-def deploy_flask(aRepo, aEntryPoint='app:app'):
+def deploy_django(aRepo, aEntryPoint='core.wsgi:application'):
     """
     Referance: https://api-docs.render.com/reference/create-service
     """
@@ -52,7 +52,7 @@ def deploy_flask(aRepo, aEntryPoint='app:app'):
     try:
 
         ownerId      = get_owner()
-        service_name = nameFromRepo( aRepo ) + '-' + randStr() 
+        service_name = nameFromRepo( aRepo ) # + '-' + randStr() 
 
         if not ownerId:
             raise Exception( 'Error getting owner' )
@@ -63,6 +63,65 @@ def deploy_flask(aRepo, aEntryPoint='app:app'):
                 {
                     "key": "DEBUG",
                     "value": "True"
+                }
+            ],            
+            'serviceDetails': {
+                'env':  'python',
+                "envSpecificDetails":{
+                    "buildCommand":"pip install --upgrade pip ; pip install -r requirements.txt; python manage.py migrate",
+                    "startCommand":f"gunicorn {aEntryPoint}"
+                },
+            },
+            'type': 'web_service',
+            'name': service_name,
+            'ownerId': ownerId,
+            'repo': aRepo,
+        }
+
+        response = requests.post(url, json=payload, headers=HEADERS)
+
+        # HTTP 201 = Resource Created
+        if 201 != response.status_code:
+            raise Exception( response.text )
+
+        response_json = json.loads( response.text )
+
+        if DEBUG:
+            print( ' > RESPONSE ' + str( response_json ) )        
+
+        deploy_id  = response_json["deployId"]
+        deploy_url = response_json["service"]["serviceDetails"]["url"]
+        
+        #if DEBUG:
+        print(" > Deploy ID ["+deploy_id+"] -> " + deploy_url)
+
+        return json.loads( response.text )
+
+    except Exception as e:
+        print(e)
+        return None 
+
+def deploy_flask(aRepo, aEntryPoint='app:app'):
+    """
+    Referance: https://api-docs.render.com/reference/create-service
+    """
+
+    url     = f"{URL}/v1/services"
+
+    try:
+
+        ownerId      = get_owner()
+        service_name = nameFromRepo( aRepo ) # + '-' + randStr() 
+
+        if not ownerId:
+            raise Exception( 'Error getting owner' )
+
+        payload = {
+            'autoDeploy': 'yes',
+            'envVars': [
+                {
+                    "key": "DEBUG",
+                    "value": "1"
                 }
             ],            
             'serviceDetails': {
@@ -85,6 +144,9 @@ def deploy_flask(aRepo, aEntryPoint='app:app'):
             raise Exception( response.text )
 
         response_json = json.loads( response.text )
+
+        if DEBUG:
+            print( ' > RESPONSE ' + str( response_json ) )        
 
         deploy_id  = response_json["deployId"]
         deploy_url = response_json["service"]["serviceDetails"]["url"]
@@ -142,6 +204,9 @@ def deploy_static(aRepo, aNodeVer=NODE_14):
             raise Exception( response.text )
 
         response_json = json.loads( response.text )
+
+        if DEBUG:
+            print( ' > RESPONSE ' + str( response_json ) )        
 
         deploy_id  = response_json["deployId"]
         deploy_url = response_json["service"]["serviceDetails"]["url"]